+++ Genesis/src/utopia/genesis/handling/template/DeepHandler.scala
@@ -3,8 +3,7 @@ package utopia.genesis.handling.template
 import utopia.flow.collection.immutable.Empty
 import utopia.flow.util.logging.Logger
 import utopia.flow.view.immutable.eventful.AlwaysTrue
-import utopia.flow.view.template.eventful.Flag._
-import utopia.flow.view.template.eventful.{Changing, Flag}
+import utopia.flow.view.template.eventful.Flag
 
 /**
   * An abstract Handler implementation which extends the Handleable trait
@@ -12,11 +11,11 @@ import utopia.flow.view.template.eventful.{Changing, Flag}
   * @since 01/02/2024, v4.0
   */
 abstract class DeepHandler[A <: Handleable](initialItems: IterableOnce[A] = Empty,
-                                            additionalHandleCondition: Changing[Boolean] = AlwaysTrue)
+                                            additionalHandleCondition: Flag = AlwaysTrue)
                                            (implicit log: Logger)
 	extends AbstractHandler[A](initialItems) with Handleable
 {
 	// ATTRIBUTES   --------------------
 	
-	override val handleCondition: Flag = itemsPointer.readOnly.map { _.nonEmpty } && additionalHandleCondition
+	override val handleCondition: Flag = itemsPointer.nonEmptyFlag && additionalHandleCondition
 }